Reorgs `fiori-generator-shared` to promote easier discovery and reuse
