+++ packages/fiori-generator-shared/src/helpers.ts
@@ -1,6 +1,3 @@
-import { existsSync } from 'fs';
-import { DEFAULT_PROJECTS_FOLDER, YEOMANUI_TARGET_FOLDER_CONFIG_PROP } from './constants';
-
 /**
  * Get the resource URLs for the UShell bootstrap and UI5 bootstrap based on project type and UI5 framework details.
  *
@@ -31,30 +28,4 @@ export function getBootstrapResourceUrls(
         isEdmxProjectType || !frameworkUrl ? `..${relativeUiPath}` : `${frameworkUrl}${versionPath}${relativeUiPath}`;
 
     return { uShellBootstrapResourceUrl, uiBootstrapResourceUrl };
-}
-
-/**
- * Determines the target folder for the project.
- *
- * @param vscode - the vscode instance
- * @returns The default path, if it can be determined otherwise undefined.
- */
-export function getDefaultTargetFolder(vscode: any): string | undefined {
-    // CLI use will not define vscode
-    if (!vscode) {
-        return undefined;
-    }
-    const targetFolder = vscode.workspace?.getConfiguration().get(YEOMANUI_TARGET_FOLDER_CONFIG_PROP);
-
-    if (targetFolder) {
-        return targetFolder;
-    }
-    const workspace = vscode.workspace;
-
-    // If this is not a workspace default to the first folder (`rootPath` is deprecated)
-    if (workspace.workspaceFolders?.length > 0 && workspace.workspaceFolders[0].uri.scheme === 'file') {
-        return workspace.workspaceFolders[0].uri.fsPath;
-    }
-    // Otherwise use <home-dir>/projects,
-    return existsSync(DEFAULT_PROJECTS_FOLDER) ? DEFAULT_PROJECTS_FOLDER : undefined;
 }
